Move the metrics report into a dedicated xla_model API. Remove the old ParallelLoader, and move the new one into the proper file.
