+++ test/test_train_mnist.py
@@ -20,6 +20,7 @@ import torch.optim as optim
 from torchvision import datasets, transforms
 import torch_xla
 import torch_xla.distributed.data_parallel as dp
+import torch_xla.debug.metrics as met
 import torch_xla.utils.utils as xu
 import torch_xla.core.xla_model as xm
 import unittest
@@ -149,7 +150,7 @@ def train_mnist():
     accuracies = model_parallel(test_loop_fn, test_loader)
     accuracy = sum(accuracies) / len(accuracies)
     if FLAGS.metrics_debug:
-      print(torch_xla._XLAC._xla_metrics_report())
+      print(met.metrics_report())
 
   return accuracy
